refactor: CoapMethodMask is a bitmask -> struct of static constexpr, not enum class
A bitmask is OR'd together, so enum class was the wrong tool: every use
carried a (uint8_t) cast at the | just to make the scoped enum compile,
which obfuscates a plain bit-OR. Make it a namespacing struct of
static constexpr uint8_t members instead - CoapMethodMask::COAP_ALLOW_GET
| CoapMethodMask::COAP_ALLOW_POST reads as the bit-OR it is, with no
casts. Values still bind to the CoapMethod ordinal (bit = 1 << method).
native_coap+coap_observe+pentest = 138 green.

+++ examples/L7-Application/27.CoapObserve/27.CoapObserve.ino
@@ -57,7 +57,7 @@ void setup()
     WiFi.setSleep(false);
 
     coap_server_init();
-    coap_server_add_resource("/count", (uint8_t)CoapMethodMask::COAP_ALLOW_GET, h_count);
+    coap_server_add_resource("/count", CoapMethodMask::COAP_ALLOW_GET, h_count);
     coap_server_begin_udp(5683);
     Serial.println("CoAP server on :5683, observe coap://<ip>/count");
 }
